backend/requirements.txt
langchain>=1.2.0
langchain-community>=0.4.0
chromadb>=1.5.0
tiktoken>=0.12.0
google-generativeai>=0.8.0
